# apps/backend-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.APIResponse:
    properties:
      count:
        type: integer
      data: {}
      error:
        type: string
      has_more:
        type: boolean
      next_cursor:
        type: string
      success:
        type: boolean
    type: object
...
  models.RegisterRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
      website:
        description: Honeypot field — must be empty
        type: string
    type: object
...
info:
  contact:
    name: gomo6 Team
    url: https://gomo6.wtf
  description: gomo6 — open social platform API. Create bots, integrations, and apps.
  license:
    name: MIT
  termsOfService: https://gomo6.wtf/terms
  title: gomo6 API
  version: "1.0"
paths:
  /auth/2fa/disable:
    post:
      description: Disable 2FA for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Disable TOTP
      tags:
      - Auth
  /auth/2fa/setup:
    post:
      description: Generate a new TOTP secret and provisioning URI for 2FA setup
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Setup TOTP
      tags:
      - Auth
  /auth/2fa/status:
    get:
      description: Returns whether 2FA is enabled and if there is a pending setup
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Get 2FA status
      tags:
      - Auth
  /auth/2fa/verify-and-enable:
    post:
      consumes:
      - application/json
      description: Verify the TOTP code and enable 2FA for the authenticated user
      parameters:
      - description: TOTP verification code
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      security:
      - BearerAuth: []
      summary: Verify and enable TOTP
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with username and password. Returns tokens or partial
        token if 2FA is enabled.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Log in
      tags:
      - Auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new account with username, password, and optional email
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Register a new user
      tags:
      - Auth
  /auth/verify-2fa:
    post:
      consumes:
      - application/json
      description: Validate a TOTP code during login (requires partial token)
      parameters:
      - description: 2FA verification request
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.APIResponse'
      summary: Verify 2FA code
      tags:
      - Auth
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Bearer token for API authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"